atualizaçoes

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: jefernan <jefernan@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/30 16:24:05 by jefernan          #+#    #+#             */
-/*   Updated: 2021/09/03 13:39:41 by jefernan         ###   ########.fr       */
+/*   Updated: 2021/09/24 11:13:10 by jefernan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (dstsize == 0)
